Split classes out of dht::Context
