+++ llarp/dht/context.cpp
@@ -1,10 +1,18 @@
 #include <dht/context.hpp>
 
+#include <dht/explorenetworkjob.hpp>
+#include <dht/localrouterlookup.hpp>
+#include <dht/localserviceaddresslookup.hpp>
+#include <dht/localtaglookup.hpp>
 #include <dht/messages/findrouter.hpp>
 #include <dht/messages/gotintro.hpp>
 #include <dht/messages/gotrouter.hpp>
 #include <dht/messages/pubintro.hpp>
 #include <dht/node.hpp>
+#include <dht/publishservicejob.hpp>
+#include <dht/recursiverouterlookup.hpp>
+#include <dht/serviceaddresslookup.hpp>
+#include <dht/taglookup.hpp>
 #include <messages/dht.hpp>
 #include <messages/dht_immediate.hpp>
 #include <router/router.hpp>
@@ -15,18 +23,9 @@ namespace llarp
 {
   namespace dht
   {
-    Context::Context()
+    Context::Context() : router(nullptr), allowTransit(false)
     {
       randombytes((byte_t *)&ids, sizeof(uint64_t));
-      allowTransit = false;
-    }
-
-    Context::~Context()
-    {
-      if(nodes)
-        delete nodes;
-      if(services)
-        delete services;
     }
 
     void
@@ -44,53 +43,6 @@ namespace llarp
       else
         llarp::LogError("failed to select random nodes for exploration");
     }
-
-    struct ExploreNetworkJob : public TX< RouterID, RouterID >
-    {
-      ExploreNetworkJob(const RouterID &peer, Context *ctx)
-          : TX< RouterID, RouterID >(TXOwner{}, peer, ctx)
-      {
-      }
-
-      bool
-      Validate(const RouterID &) const override
-      {
-        // TODO: check with lokid
-        return true;
-      }
-
-      void
-      Start(const TXOwner &peer) override
-      {
-        parent->DHTSendTo(peer.node.as_array(),
-                          new FindRouterMessage(peer.txid));
-      }
-
-      bool
-      GetNextPeer(Key_t &, const std::set< Key_t > &) override
-      {
-        return false;
-      }
-
-      void
-      DoNextRequest(const Key_t &) override
-      {
-      }
-
-      void
-      SendReply() override
-      {
-        llarp::LogInfo("got ", valuesFound.size(), " routers from exploration");
-        for(const auto &pk : valuesFound)
-        {
-          // lookup router
-          parent->LookupRouter(
-              pk,
-              std::bind(&llarp::Router::HandleDHTLookupForExplore,
-                        parent->router, pk, std::placeholders::_1));
-        }
-      }
-    };
 
     void
     Context::ExploreNetworkVia(const Key_t &askpeer)
@@ -272,8 +224,8 @@ namespace llarp
     {
       router   = r;
       ourKey   = us;
-      nodes    = new Bucket< RCNode >(ourKey, llarp::randint);
-      services = new Bucket< ISNode >(ourKey, llarp::randint);
+      nodes    = std::make_unique< Bucket< RCNode > >(ourKey, llarp::randint);
+      services = std::make_unique< Bucket< ISNode > >(ourKey, llarp::randint);
       llarp::LogDebug("initialize dht with key ", ourKey);
       // start exploring
 
@@ -308,119 +260,13 @@ namespace llarp
       llarp::routing::DHTMessage reply;
       if(!msg->HandleMessage(router->dht, reply.M))
         return false;
-      if(reply.M.size())
+      if(!reply.M.empty())
       {
         auto path = router->paths.GetByUpstream(router->pubkey(), id);
         return path && path->SendRoutingMessage(&reply, router);
       }
       return true;
     }
-
-    struct ServiceAddressLookup
-        : public TX< service::Address, service::IntroSet >
-    {
-      IntroSetLookupHandler handleResult;
-      uint64_t R;
-
-      ServiceAddressLookup(const TXOwner &asker, const service::Address &addr,
-                           Context *ctx, uint64_t r,
-                           IntroSetLookupHandler handler)
-          : TX< service::Address, service::IntroSet >(asker, addr, ctx)
-          , handleResult(handler)
-          , R(r)
-      {
-        peersAsked.insert(ctx->OurKey());
-      }
-
-      bool
-      Validate(const service::IntroSet &value) const override
-      {
-        if(!value.Verify(parent->Crypto(), parent->Now()))
-        {
-          llarp::LogWarn("Got invalid introset from service lookup");
-          return false;
-        }
-        if(value.A.Addr() != target)
-        {
-          llarp::LogWarn("got introset with wrong target from service lookup");
-          return false;
-        }
-        return true;
-      }
-
-      bool
-      GetNextPeer(Key_t &next, const std::set< Key_t > &exclude) override
-      {
-        Key_t k = target.ToKey();
-        return parent->nodes->FindCloseExcluding(k, next, exclude);
-      }
-
-      void
-      Start(const TXOwner &peer) override
-      {
-        parent->DHTSendTo(peer.node.as_array(),
-                          new FindIntroMessage(peer.txid, target, R));
-      }
-
-      void
-      DoNextRequest(const Key_t &ask) override
-      {
-        if(R)
-          parent->LookupIntroSetRecursive(target, whoasked.node, whoasked.txid,
-                                          ask, R - 1);
-        else
-          parent->LookupIntroSetIterative(target, whoasked.node, whoasked.txid,
-                                          ask);
-      }
-
-      virtual void
-      SendReply() override
-      {
-        if(handleResult)
-          handleResult(valuesFound);
-
-        parent->DHTSendTo(whoasked.node.as_array(),
-                          new GotIntroMessage(valuesFound, whoasked.txid));
-      }
-    };
-
-    struct LocalServiceAddressLookup : public ServiceAddressLookup
-    {
-      PathID_t localPath;
-
-      LocalServiceAddressLookup(const PathID_t &pathid, uint64_t txid,
-                                const service::Address &addr, Context *ctx,
-                                __attribute__((unused)) const Key_t &askpeer)
-          : ServiceAddressLookup(TXOwner{ctx->OurKey(), txid}, addr, ctx, 5,
-                                 nullptr)
-          , localPath(pathid)
-      {
-      }
-
-      void
-      SendReply() override
-      {
-        auto path = parent->router->paths.GetByUpstream(
-            parent->OurKey().as_array(), localPath);
-        if(!path)
-        {
-          llarp::LogWarn(
-              "did not send reply for relayed dht request, no such local path "
-              "for pathid=",
-              localPath);
-          return;
-        }
-        routing::DHTMessage msg;
-        msg.M.emplace_back(new GotIntroMessage(valuesFound, whoasked.txid));
-        if(!path->SendRoutingMessage(&msg, parent->router))
-        {
-          llarp::LogWarn(
-              "failed to send routing message when informing result of dht "
-              "request, pathid=",
-              localPath);
-        }
-      }
-    };
 
     void
     Context::LookupIntroSetForPath(const service::Address &addr, uint64_t txid,
@@ -433,62 +279,6 @@ namespace llarp
           peer, asker, addr,
           new LocalServiceAddressLookup(path, txid, addr, this, askpeer));
     }
-
-    struct PublishServiceJob : public TX< service::Address, service::IntroSet >
-    {
-      uint64_t S;
-      std::set< Key_t > dontTell;
-      service::IntroSet I;
-      PublishServiceJob(const TXOwner &asker, const service::IntroSet &introset,
-                        Context *ctx, uint64_t s,
-                        const std::set< Key_t > &exclude)
-          : TX< service::Address, service::IntroSet >(asker, introset.A.Addr(),
-                                                      ctx)
-          , S(s)
-          , dontTell(exclude)
-          , I(introset)
-      {
-      }
-
-      bool
-      Validate(const service::IntroSet &introset) const override
-      {
-        if(I.A != introset.A)
-        {
-          llarp::LogWarn(
-              "publish introset acknowledgement acked a different service");
-          return false;
-        }
-        return true;
-      }
-
-      void
-      Start(const TXOwner &peer) override
-      {
-        std::vector< Key_t > exclude;
-        for(const auto &router : dontTell)
-          exclude.push_back(router);
-        parent->DHTSendTo(peer.node.as_array(),
-                          new PublishIntroMessage(I, peer.txid, S, exclude));
-      }
-
-      bool
-      GetNextPeer(Key_t &, const std::set< Key_t > &) override
-      {
-        return false;
-      }
-
-      void
-      DoNextRequest(const Key_t &) override
-      {
-      }
-
-      void
-      SendReply() override
-      {
-        // don't need this
-      }
-    };
 
     void
     Context::PropagateIntroSetTo(const Key_t &from, uint64_t txid,
@@ -508,7 +298,7 @@ namespace llarp
     Context::LookupIntroSetRecursive(const service::Address &addr,
                                      const Key_t &whoasked, uint64_t txid,
                                      const Key_t &askpeer, uint64_t R,
-                                     IntroSetLookupHandler handler)
+                                     service::IntroSetLookupHandler handler)
     {
       TXOwner asker(whoasked, txid);
       TXOwner peer(askpeer, ++ids);
@@ -521,7 +311,7 @@ namespace llarp
     Context::LookupIntroSetIterative(const service::Address &addr,
                                      const Key_t &whoasked, uint64_t txid,
                                      const Key_t &askpeer,
-                                     IntroSetLookupHandler handler)
+                                     service::IntroSetLookupHandler handler)
     {
       TXOwner asker(whoasked, txid);
       TXOwner peer(askpeer, ++ids);
@@ -529,76 +319,6 @@ namespace llarp
           peer, asker, addr,
           new ServiceAddressLookup(asker, addr, this, 0, handler));
     }
-
-    struct TagLookup : public TX< service::Tag, service::IntroSet >
-    {
-      uint64_t R;
-      TagLookup(const TXOwner &asker, const service::Tag &tag, Context *ctx,
-                uint64_t r)
-          : TX< service::Tag, service::IntroSet >(asker, tag, ctx), R(r)
-      {
-      }
-
-      bool
-      Validate(const service::IntroSet &introset) const override
-      {
-        if(!introset.Verify(parent->Crypto(), parent->Now()))
-        {
-          llarp::LogWarn("got invalid introset from tag lookup");
-          return false;
-        }
-        if(introset.topic != target)
-        {
-          llarp::LogWarn("got introset with missmatched topic in tag lookup");
-          return false;
-        }
-        return true;
-      }
-
-      void
-      Start(const TXOwner &peer) override
-      {
-        parent->DHTSendTo(peer.node.as_array(),
-                          new FindIntroMessage(target, peer.txid, R));
-      }
-
-      bool
-      GetNextPeer(Key_t &, const std::set< Key_t > &) override
-      {
-        return false;
-      }
-
-      void
-      DoNextRequest(const Key_t &) override
-      {
-      }
-
-      void
-      SendReply() override
-      {
-        std::set< service::IntroSet > found;
-        for(const auto &remoteTag : valuesFound)
-        {
-          found.insert(remoteTag);
-        }
-        // collect our local values if we haven't hit a limit
-        if(found.size() < 2)
-        {
-          for(const auto &localTag :
-              parent->FindRandomIntroSetsWithTagExcluding(target, 1, found))
-          {
-            found.insert(localTag);
-          }
-        }
-        std::vector< service::IntroSet > values;
-        for(const auto &introset : found)
-        {
-          values.push_back(introset);
-        }
-        parent->DHTSendTo(whoasked.node.as_array(),
-                          new GotIntroMessage(values, whoasked.txid));
-      }
-    };
 
     void
     Context::LookupTagRecursive(const service::Tag &tag, const Key_t &whoasked,
@@ -612,42 +332,6 @@ namespace llarp
       llarp::LogInfo("ask ", askpeer, " for ", tag, " on behalf of ", whoasked,
                      " R=", R);
     }
-
-    struct LocalTagLookup : public TagLookup
-    {
-      PathID_t localPath;
-
-      LocalTagLookup(const PathID_t &path, uint64_t txid,
-                     const service::Tag &target, Context *ctx)
-          : TagLookup(TXOwner{ctx->OurKey(), txid}, target, ctx, 0)
-          , localPath(path)
-      {
-      }
-
-      void
-      SendReply() override
-      {
-        auto path = parent->router->paths.GetByUpstream(
-            parent->OurKey().as_array(), localPath);
-        if(!path)
-        {
-          llarp::LogWarn(
-              "did not send reply for relayed dht request, no such local path "
-              "for pathid=",
-              localPath);
-          return;
-        }
-        routing::DHTMessage msg;
-        msg.M.emplace_back(new GotIntroMessage(valuesFound, whoasked.txid));
-        if(!path->SendRoutingMessage(&msg, parent->router))
-        {
-          llarp::LogWarn(
-              "failed to send routing message when informing result of dht "
-              "request, pathid=",
-              localPath);
-        }
-      }
-    };
 
     void
     Context::LookupTagForPath(const service::Tag &tag, uint64_t txid,
@@ -698,101 +382,6 @@ namespace llarp
       reply.emplace_back(new GotRouterMessage(txid, closer, false));
       return true;
     }
-
-    struct RecursiveRouterLookup : public TX< RouterID, RouterContact >
-    {
-      RouterLookupHandler resultHandler;
-      RecursiveRouterLookup(const TXOwner &whoasked, const RouterID &target,
-                            Context *ctx, RouterLookupHandler result)
-          : TX< RouterID, RouterContact >(whoasked, target, ctx)
-          , resultHandler(result)
-
-      {
-        peersAsked.insert(ctx->OurKey());
-      }
-
-      bool
-      Validate(const RouterContact &rc) const override
-      {
-        if(!rc.Verify(parent->Crypto(), parent->Now()))
-        {
-          llarp::LogWarn("rc from lookup result is invalid");
-          return false;
-        }
-        return true;
-      }
-
-      bool
-      GetNextPeer(Key_t &, const std::set< Key_t > &) override
-      {
-        return false;
-      }
-
-      void
-      DoNextRequest(const Key_t &) override
-      {
-      }
-
-      void
-      Start(const TXOwner &peer) override
-      {
-        parent->DHTSendTo(peer.node.as_array(),
-                          new FindRouterMessage(peer.txid, target));
-      }
-
-      virtual void
-      SendReply() override
-      {
-        if(resultHandler)
-        {
-          resultHandler(valuesFound);
-        }
-        else
-        {
-          parent->DHTSendTo(
-              whoasked.node.as_array(),
-              new GotRouterMessage({}, whoasked.txid, valuesFound, false));
-        }
-      }
-    };
-
-    struct LocalRouterLookup : public RecursiveRouterLookup
-    {
-      PathID_t localPath;
-
-      LocalRouterLookup(const PathID_t &path, uint64_t txid,
-                        const RouterID &target, Context *ctx)
-          : RecursiveRouterLookup(TXOwner{ctx->OurKey(), txid}, target, ctx,
-                                  nullptr)
-          , localPath(path)
-      {
-      }
-
-      void
-      SendReply() override
-      {
-        auto path = parent->router->paths.GetByUpstream(
-            parent->OurKey().as_array(), localPath);
-        if(!path)
-        {
-          llarp::LogWarn(
-              "did not send reply for relayed dht request, no such local path "
-              "for pathid=",
-              localPath);
-          return;
-        }
-        routing::DHTMessage msg;
-        msg.M.emplace_back(new GotRouterMessage(parent->OurKey(), whoasked.txid,
-                                                valuesFound, true));
-        if(!path->SendRoutingMessage(&msg, parent->router))
-        {
-          llarp::LogWarn(
-              "failed to send routing message when informing result of dht "
-              "request, pathid=",
-              localPath);
-        }
-      }
-    };
 
     void
     Context::LookupRouterForPath(const RouterID &target, uint64_t txid,
